add a combined test executable
