+++ test/object-store/CMakeLists.txt
@@ -1,7 +1,4 @@
 set(HEADERS
-    ../util/crypt_key.hpp
-    ../util/spawned_process.hpp
-    ../util/test_path.hpp
     util/event_loop.hpp
     util/index_helpers.hpp
     util/test_file.hpp
@@ -10,13 +7,15 @@ set(HEADERS
 )
 
 set(SOURCES
+    # slowest to compile first
+    set.cpp
+
     backup.cpp
     collection_change_indices.cpp
     dictionary.cpp
     frozen_objects.cpp
     index_set.cpp
     list.cpp
-    main.cpp
     migrations.cpp
     object.cpp
     object_store.cpp
@@ -25,16 +24,13 @@ set(SOURCES
     results.cpp
     schema.cpp
     sectioned_results.cpp
-    set.cpp
+    test_runner.cpp
     thread_safe_reference.cpp
     transaction_log_parsing.cpp
     uuid.cpp
     c_api/c_api.cpp
-    c_api/c_api.c
+    c_api/c_api_file_tests.c
 
-    ../util/crypt_key.cpp
-    ../util/spawned_process.cpp
-    ../util/test_path.cpp
     util/event_loop.cpp
     util/test_file.cpp
     util/test_utils.cpp
@@ -44,9 +40,7 @@ if (REALM_ENABLE_GEOSPATIAL)
     list(APPEND SOURCES geospatial.cpp)
 endif()
 
-file(GLOB RESOURCES RELATIVE ${CMAKE_CURRENT_BINARY_DIR}
-    ../*.pem
-     *.realm)
+file(GLOB RESOURCES "*.realm" "../*.pem")
 
 if(REALM_ENABLE_AUTH_TESTS)
     list(APPEND SOURCES util/sync/baas_admin_api.cpp)
@@ -82,16 +76,23 @@ if(REALM_ENABLE_SYNC)
     endif()
 endif()
 
-add_executable(ObjectStoreTests ${SOURCES} ${HEADERS} ${RESOURCES})
-set_target_properties(ObjectStoreTests PROPERTIES OUTPUT_NAME realm-object-store-tests)
-set_target_resources(ObjectStoreTests "${RESOURCES}")
+set_property(DIRECTORY PROPERTY TEST_RESOURCES "${RESOURCES}")
+
+add_library(ObjectStoreTestLib OBJECT ${SOURCES} ${HEADERS} ${RESOURCES})
 
 if(MSVC)
     # increase the number of sections supported in an obj file for the heavily templated tests
-    target_compile_options(ObjectStoreTests PRIVATE /bigobj)
+    target_compile_options(ObjectStoreTestLib PRIVATE /bigobj)
 endif()
 
-target_link_libraries(ObjectStoreTests Catch2::Catch2 ObjectStore RealmFFIStatic)
+target_link_libraries(ObjectStoreTestLib Catch2::Catch2 ObjectStore RealmFFIStatic TestUtil)
+enable_stdfilesystem(ObjectStoreTestLib)
+
+add_executable(ObjectStoreTests main.cpp ${RESOURCES})
+set_target_properties(ObjectStoreTests PROPERTIES OUTPUT_NAME realm-object-store-tests)
+target_sources(ObjectStoreTests PUBLIC $<TARGET_OBJECTS:ObjectStoreTestLib>)
+target_link_libraries(ObjectStoreTests ObjectStoreTestLib TestUtil)
+set_target_resources(ObjectStoreTests "${RESOURCES}")
 enable_stdfilesystem(ObjectStoreTests)
 create_coverage_target(generate-coverage ObjectStoreTests)
 
@@ -124,31 +125,31 @@ if(REALM_ENABLE_SYNC)
 endif()
 
 if(REALM_ENABLE_SYNC)
-    target_link_libraries(ObjectStoreTests SyncServer)
+    target_link_libraries(ObjectStoreTestLib SyncServer)
     option(REALM_ENABLE_AUTH_TESTS "" OFF)
     if(REALM_ENABLE_AUTH_TESTS)
         if(NOT REALM_MONGODB_ENDPOINT)
             message(FATAL_ERROR "REALM_MONGODB_ENDPOINT must be set when specifying REALM_ENABLE_AUTH_TESTS.")
         endif()
 
-        target_compile_definitions(ObjectStoreTests PRIVATE
+        target_compile_definitions(ObjectStoreTestLib PRIVATE
             REALM_ENABLE_AUTH_TESTS=1
             REALM_MONGODB_ENDPOINT="${REALM_MONGODB_ENDPOINT}"
         )
 
         find_package(CURL REQUIRED)
-        target_link_libraries(ObjectStoreTests CURL::libcurl)
+        target_link_libraries(ObjectStoreTestLib CURL::libcurl)
     endif()
 endif()
 
 if(REALM_TEST_LOGGING)
-    target_compile_definitions(ObjectStoreTests PRIVATE
+    target_compile_definitions(ObjectStoreTestLib PRIVATE
         TEST_ENABLE_LOGGING=1
     )
 
     if(REALM_TEST_LOGGING_LEVEL)
         message(STATUS "Test logging level: ${REALM_TEST_LOGGING_LEVEL}")
-        target_compile_definitions(ObjectStoreTests PRIVATE
+        target_compile_definitions(ObjectStoreTestLib PRIVATE
             TEST_LOGGING_LEVEL=${REALM_TEST_LOGGING_LEVEL}
         )
     else()
@@ -156,7 +157,7 @@ if(REALM_TEST_LOGGING)
     endif()
 endif()
 
-target_include_directories(ObjectStoreTests PRIVATE
+target_include_directories(ObjectStoreTestLib PRIVATE
     ${CATCH_INCLUDE_DIR}
     ${JSON_INCLUDE_DIR}
     ${CMAKE_CURRENT_SOURCE_DIR}
@@ -165,7 +166,7 @@ target_include_directories(ObjectStoreTests PRIVATE
 # on Apple platforms we use the built-in CFRunLoop
 # on WebAssembly we use an Emscripten-specific Scheduler and runloop
 # everywhere else it's libuv, except UWP where it doesn't build
-if(NOT APPLE AND NOT EMSCRIPTEN AND NOT WINDOWS_STORE)
+if(NOT APPLE AND NOT EMSCRIPTEN AND NOT WINDOWS_STORE AND NOT ANDROID)
     option(REALM_FETCH_MISSING_DEPENDENCIES "Download missing dependencies with CMake's FetchContent where possible" ON)
     if(REALM_FETCH_MISSING_DEPENDENCIES)
         find_package(LibUV)
@@ -192,7 +193,7 @@ if(NOT APPLE AND NOT EMSCRIPTEN AND NOT WINDOWS_STORE)
         set(libuv_target uv_a)
     endif()
 
-    target_link_libraries(ObjectStoreTests ${libuv_target})
+    target_link_libraries(ObjectStoreTestLib ${libuv_target})
     # FIXME: ObjectStore itself shouldn't care about this, but we need to refactor scheduler.cpp to make it happen
     target_compile_definitions(ObjectStore PUBLIC REALM_HAVE_UV=1)
     get_property(libuv_include_dir TARGET ${libuv_target} PROPERTY INCLUDE_DIRECTORIES)
