+++ test/test_sync.cpp
@@ -127,6 +127,7 @@ ClientHistory& get_history(DBRef db)
     return get_replication(db).get_history();
 }
 
+#if !REALM_MOBILE // the server is not implemented on devices
 TEST(Sync_BadVirtualPath)
 {
     // NOTE:  This test is no longer valid after migration to MongoDB Realm
@@ -5547,50 +5548,6 @@ TEST(Sync_ServerSideEncryption)
     CHECK(group.has_table("class_Test"));
 }
 
-
-// This test calls row_for_object_id() for various object ids and tests that
-// the right value is returned including that no assertions are hit.
-TEST(Sync_RowForGlobalKey)
-{
-    TEST_CLIENT_DB(db);
-
-    {
-        WriteTransaction wt(db);
-        TableRef table = wt.add_table("class_foo");
-        table->add_column(type_Int, "i");
-        wt.commit();
-    }
-
-    // Check that various object_ids are not in the table.
-    {
-        ReadTransaction rt(db);
-        ConstTableRef table = rt.get_table("class_foo");
-        CHECK(table);
-
-        // Default constructed GlobalKey
-        {
-            GlobalKey object_id;
-            auto row_ndx = table->get_objkey(object_id);
-            CHECK_NOT(row_ndx);
-        }
-
-        // GlobalKey with small lo and hi values
-        {
-            GlobalKey object_id{12, 24};
-            auto row_ndx = table->get_objkey(object_id);
-            CHECK_NOT(row_ndx);
-        }
-
-        // GlobalKey with lo and hi values past the 32 bit limit.
-        {
-            GlobalKey object_id{uint_fast64_t(1) << 50, uint_fast64_t(1) << 52};
-            auto row_ndx = table->get_objkey(object_id);
-            CHECK_NOT(row_ndx);
-        }
-    }
-}
-
-
 TEST(Sync_LogCompaction_EraseObject_LinkList)
 {
     TEST_DIR(dir);
@@ -6322,89 +6279,6 @@ TEST(Sync_Set)
     }
 }
 
-TEST(Sync_DanglingLinksCountInPriorSize)
-{
-    SHARED_GROUP_TEST_PATH(path);
-    ClientReplication repl;
-    auto local_db = realm::DB::create(repl, path);
-    auto& history = repl.get_history();
-    history.set_client_file_ident(sync::SaltedFileIdent{1, 123456}, true);
-
-    version_type last_version, last_version_observed = 0;
-    auto dump_uploadable = [&] {
-        UploadCursor upload_cursor{last_version_observed, 0};
-        std::vector<sync::ClientHistory::UploadChangeset> changesets_to_upload;
-        version_type locked_server_version = 0;
-        history.find_uploadable_changesets(upload_cursor, last_version, changesets_to_upload, locked_server_version);
-        CHECK_EQUAL(changesets_to_upload.size(), static_cast<size_t>(1));
-        realm::sync::Changeset parsed_changeset;
-        auto unparsed_changeset = changesets_to_upload[0].changeset.get_first_chunk();
-        realm::util::SimpleInputStream changeset_stream(unparsed_changeset);
-        realm::sync::parse_changeset(changeset_stream, parsed_changeset);
-        test_context.logger->info("changeset at version %1: %2", last_version, parsed_changeset);
-        last_version_observed = last_version;
-        return parsed_changeset;
-    };
-
-    TableKey source_table_key, target_table_key;
-    {
-        auto wt = local_db->start_write();
-        auto source_table = wt->add_table_with_primary_key("class_source", type_String, "_id");
-        auto target_table = wt->add_table_with_primary_key("class_target", type_String, "_id");
-        source_table->add_column_list(*target_table, "links");
-
-        source_table_key = source_table->get_key();
-        target_table_key = target_table->get_key();
-
-        auto obj_to_keep = target_table->create_object_with_primary_key(std::string{"target1"});
-        auto obj_to_delete = target_table->create_object_with_primary_key(std::string{"target2"});
-        auto source_obj = source_table->create_object_with_primary_key(std::string{"source"});
-
-        auto links_list = source_obj.get_linklist("links");
-        links_list.add(obj_to_keep.get_key());
-        links_list.add(obj_to_delete.get_key());
-        last_version = wt->commit();
-    }
-
-    dump_uploadable();
-
-    {
-        // Simulate removing the object via the sync client so we get a dangling link
-        TempShortCircuitReplication disable_repl(repl);
-        auto wt = local_db->start_write();
-        auto target_table = wt->get_table(target_table_key);
-        auto obj = target_table->get_object_with_primary_key(std::string{"target2"});
-        obj.invalidate();
-        last_version = wt->commit();
-    }
-
-    {
-        auto wt = local_db->start_write();
-        auto source_table = wt->get_table(source_table_key);
-        auto target_table = wt->get_table(target_table_key);
-
-        auto obj_to_add = target_table->create_object_with_primary_key(std::string{"target3"});
-
-        auto source_obj = source_table->get_object_with_primary_key(std::string{"source"});
-        auto links_list = source_obj.get_linklist("links");
-        links_list.add(obj_to_add.get_key());
-        last_version = wt->commit();
-    }
-
-    auto changeset = dump_uploadable();
-    CHECK_EQUAL(changeset.size(), static_cast<size_t>(2));
-    auto changeset_it = changeset.end();
-    --changeset_it;
-    auto last_instr = *changeset_it;
-    CHECK_EQUAL(last_instr->type(), Instruction::Type::ArrayInsert);
-    auto arr_insert_instr = last_instr->get_as<Instruction::ArrayInsert>();
-    CHECK_EQUAL(changeset.get_string(arr_insert_instr.table), StringData("source"));
-    CHECK(arr_insert_instr.value.type == sync::instr::Payload::Type::Link);
-    CHECK_EQUAL(changeset.get_string(mpark::get<InternString>(arr_insert_instr.value.data.link.target)),
-                StringData("target3"));
-    CHECK_EQUAL(arr_insert_instr.prior_size, 2);
-}
-
 TEST(Sync_BundledRealmFile)
 {
     TEST_CLIENT_DB(db);
@@ -6589,109 +6463,6 @@ TEST(Sync_MergeStringPrimaryKey)
     }
 }
 
-TEST(Sync_NonIncreasingServerVersions)
-{
-    TEST_CLIENT_DB(db);
-
-    auto& history = get_history(db);
-    history.set_client_file_ident(SaltedFileIdent{2, 0x1234567812345678}, false);
-    timestamp_type timestamp{1};
-    history.set_local_origin_timestamp_source([&] {
-        return ++timestamp;
-    });
-
-    auto latest_local_version = [&] {
-        auto tr = db->start_write();
-        tr->add_table_with_primary_key("class_foo", type_String, "_id")->add_column(type_Int, "int_col");
-        return tr->commit();
-    }();
-
-    std::vector<Changeset> server_changesets;
-    auto prep_changeset = [&](auto pk_name, auto int_col_val) {
-        Changeset changeset;
-        changeset.version = 10;
-        changeset.last_integrated_remote_version = latest_local_version - 1;
-        changeset.origin_timestamp = ++timestamp;
-        changeset.origin_file_ident = 1;
-        instr::PrimaryKey pk{changeset.intern_string(pk_name)};
-        auto table_name = changeset.intern_string("foo");
-        auto col_name = changeset.intern_string("int_col");
-        instr::EraseObject erase_1;
-        erase_1.object = pk;
-        erase_1.table = table_name;
-        changeset.push_back(erase_1);
-        instr::CreateObject create_1;
-        create_1.object = pk;
-        create_1.table = table_name;
-        changeset.push_back(create_1);
-        instr::Update update_1;
-        update_1.table = table_name;
-        update_1.object = pk;
-        update_1.field = col_name;
-        update_1.value = instr::Payload{int64_t(int_col_val)};
-        changeset.push_back(update_1);
-        server_changesets.push_back(std::move(changeset));
-    };
-    prep_changeset("bizz", 1);
-    prep_changeset("buzz", 2);
-    prep_changeset("baz", 3);
-    prep_changeset("bar", 4);
-    ++server_changesets.back().version;
-
-    std::vector<ChangesetEncoder::Buffer> encoded;
-    std::vector<Transformer::RemoteChangeset> server_changesets_encoded;
-    for (const auto& changeset : server_changesets) {
-        encoded.emplace_back();
-        encode_changeset(changeset, encoded.back());
-        server_changesets_encoded.emplace_back(changeset.version, changeset.last_integrated_remote_version,
-                                               BinaryData(encoded.back().data(), encoded.back().size()),
-                                               changeset.origin_timestamp, changeset.origin_file_ident);
-    }
-
-    SyncProgress progress = {};
-    progress.download.server_version = server_changesets.back().version;
-    progress.download.last_integrated_client_version = latest_local_version - 1;
-    progress.latest_server_version.version = server_changesets.back().version;
-    progress.latest_server_version.salt = 0x7876543217654321;
-
-    uint_fast64_t downloadable_bytes = 0;
-    VersionInfo version_info;
-    util::StderrLogger logger;
-    auto transact = db->start_read();
-    history.integrate_server_changesets(progress, &downloadable_bytes, server_changesets_encoded, version_info,
-                                        DownloadBatchState::SteadyState, logger, transact);
-}
-
-TEST(Sync_InvalidChangesetFromServer)
-{
-    TEST_CLIENT_DB(db);
-
-    auto& history = get_history(db);
-    history.set_client_file_ident(SaltedFileIdent{2, 0x1234567812345678}, false);
-
-    instr::CreateObject bad_instr;
-    bad_instr.object = InternString{1};
-    bad_instr.table = InternString{2};
-
-    Changeset changeset;
-    changeset.push_back(bad_instr);
-
-    ChangesetEncoder::Buffer encoded;
-    encode_changeset(changeset, encoded);
-    Transformer::RemoteChangeset server_changeset;
-    server_changeset.origin_file_ident = 1;
-    server_changeset.remote_version = 1;
-    server_changeset.data = BinaryData(encoded.data(), encoded.size());
-
-    VersionInfo version_info;
-    util::StderrLogger logger;
-    auto transact = db->start_read();
-    CHECK_THROW_EX(history.integrate_server_changesets({}, nullptr, util::Span(&server_changeset, 1), version_info,
-                                                       DownloadBatchState::SteadyState, logger, transact),
-                   sync::IntegrationException,
-                   StringData(e.what()).contains("Failed to parse received changeset: Invalid interned string"));
-}
-
 TEST(Sync_DifferentUsersMultiplexing)
 {
     ClientServerFixture::Config fixture_config;
@@ -6734,6 +6505,92 @@ TEST(Sync_DifferentUsersMultiplexing)
     CHECK_NOT_EQUAL(user_1_sess_2.sess.get_appservices_connection_id(),
                     user_2_sess_2.sess.get_appservices_connection_id());
 }
+
+TEST(Sync_TransformAgainstEmptyReciprocalChangeset)
+{
+    TEST_CLIENT_DB(seed_db);
+    TEST_CLIENT_DB(db_1);
+    TEST_CLIENT_DB(db_2);
+
+    {
+        auto tr = seed_db->start_write();
+        // Create schema: single table with array of ints as property.
+        auto table = tr->add_table_with_primary_key("class_table", type_Int, "_id");
+        table->add_column_list(type_Int, "ints");
+        table->add_column(type_String, "string");
+        tr->commit_and_continue_writing();
+
+        // Create object and initialize array.
+        table = tr->get_table("class_table");
+        auto obj = table->create_object_with_primary_key(42);
+        auto ints = obj.get_list<int64_t>("ints");
+        for (auto i = 0; i < 8; ++i) {
+            ints.insert(i, i);
+        }
+        tr->commit();
+    }
+
+    {
+        TEST_DIR(dir);
+        MultiClientServerFixture fixture(3, 1, dir, test_context);
+        fixture.start();
+
+        util::Optional<Session> seed_session = fixture.make_bound_session(0, seed_db, 0, "/test");
+        util::Optional<Session> db_1_session = fixture.make_bound_session(1, db_1, 0, "/test");
+        util::Optional<Session> db_2_session = fixture.make_bound_session(2, db_2, 0, "/test");
+
+        seed_session->wait_for_upload_complete_or_client_stopped();
+        db_1_session->wait_for_download_complete_or_client_stopped();
+        db_2_session->wait_for_download_complete_or_client_stopped();
+        seed_session.reset();
+        db_2_session.reset();
+
+        auto move_element = [&](const DBRef& db, size_t from, size_t to, size_t string_size = 0) {
+            auto wt = db->start_write();
+            auto table = wt->get_table("class_table");
+            auto obj = table->get_object_with_primary_key(42);
+            auto ints = obj.get_list<int64_t>("ints");
+            ints.move(from, to);
+            obj.set("string", std::string(string_size, 'a'));
+            wt->commit();
+        };
+
+        // Client 1 uploads two move instructions.
+        move_element(db_1, 7, 2);
+        move_element(db_1, 7, 6);
+
+        db_1_session->wait_for_upload_complete_or_client_stopped();
+
+        std::this_thread::sleep_for(std::chrono::milliseconds{10});
+
+        // Client 2 uploads two move instructions.
+        // The sync client uploads at most 128 KB of data so we make the first changeset large enough so two upload
+        // messages are sent to the server instead of one. Each change is transformed against the changes from
+        // Client 1.
+
+        // First change discards the first change (move(7, 2)) of Client 1.
+        move_element(db_2, 7, 0, 200 * 1024);
+        // Second change is tranformed against an empty reciprocal changeset as result of the change above.
+        move_element(db_2, 7, 5);
+        db_2_session = fixture.make_bound_session(2, db_2, 0, "/test");
+
+        db_1_session->wait_for_upload_complete_or_client_stopped();
+        db_2_session->wait_for_upload_complete_or_client_stopped();
+
+        db_1_session->wait_for_download_complete_or_client_stopped();
+        db_2_session->wait_for_download_complete_or_client_stopped();
+    }
+
+    ReadTransaction rt_1(db_1);
+    ReadTransaction rt_2(db_2);
+    const Group& group_1 = rt_1;
+    const Group& group_2 = rt_2;
+    group_1.verify();
+    group_2.verify();
+    CHECK(compare_groups(rt_1, rt_2));
+}
+
+#endif // !REALM_MOBILE
 
 // Tests that an empty reciprocal changesets is set and retrieved correctly.
 TEST(Sync_SetAndGetEmptyReciprocalChangeset)
@@ -6825,88 +6682,35 @@ TEST(Sync_SetAndGetEmptyReciprocalChangeset)
     CHECK(reciprocal_changeset.empty());
 }
 
-TEST(Sync_TransformAgainstEmptyReciprocalChangeset)
+
+TEST(Sync_InvalidChangesetFromServer)
 {
-    TEST_CLIENT_DB(seed_db);
-    TEST_CLIENT_DB(db_1);
-    TEST_CLIENT_DB(db_2);
+    TEST_CLIENT_DB(db);
 
-    {
-        auto tr = seed_db->start_write();
-        // Create schema: single table with array of ints as property.
-        auto table = tr->add_table_with_primary_key("class_table", type_Int, "_id");
-        table->add_column_list(type_Int, "ints");
-        table->add_column(type_String, "string");
-        tr->commit_and_continue_writing();
+    auto& history = get_history(db);
+    history.set_client_file_ident(SaltedFileIdent{2, 0x1234567812345678}, false);
 
-        // Create object and initialize array.
-        table = tr->get_table("class_table");
-        auto obj = table->create_object_with_primary_key(42);
-        auto ints = obj.get_list<int64_t>("ints");
-        for (auto i = 0; i < 8; ++i) {
-            ints.insert(i, i);
-        }
-        tr->commit();
-    }
+    instr::CreateObject bad_instr;
+    bad_instr.object = InternString{1};
+    bad_instr.table = InternString{2};
 
-    {
-        TEST_DIR(dir);
-        MultiClientServerFixture fixture(3, 1, dir, test_context);
-        fixture.start();
+    Changeset changeset;
+    changeset.push_back(bad_instr);
 
-        util::Optional<Session> seed_session = fixture.make_bound_session(0, seed_db, 0, "/test");
-        util::Optional<Session> db_1_session = fixture.make_bound_session(1, db_1, 0, "/test");
-        util::Optional<Session> db_2_session = fixture.make_bound_session(2, db_2, 0, "/test");
+    ChangesetEncoder::Buffer encoded;
+    encode_changeset(changeset, encoded);
+    Transformer::RemoteChangeset server_changeset;
+    server_changeset.origin_file_ident = 1;
+    server_changeset.remote_version = 1;
+    server_changeset.data = BinaryData(encoded.data(), encoded.size());
 
-        seed_session->wait_for_upload_complete_or_client_stopped();
-        db_1_session->wait_for_download_complete_or_client_stopped();
-        db_2_session->wait_for_download_complete_or_client_stopped();
-        seed_session.reset();
-        db_2_session.reset();
-
-        auto move_element = [&](const DBRef& db, size_t from, size_t to, size_t string_size = 0) {
-            auto wt = db->start_write();
-            auto table = wt->get_table("class_table");
-            auto obj = table->get_object_with_primary_key(42);
-            auto ints = obj.get_list<int64_t>("ints");
-            ints.move(from, to);
-            obj.set("string", std::string(string_size, 'a'));
-            wt->commit();
-        };
-
-        // Client 1 uploads two move instructions.
-        move_element(db_1, 7, 2);
-        move_element(db_1, 7, 6);
-
-        db_1_session->wait_for_upload_complete_or_client_stopped();
-
-        std::this_thread::sleep_for(std::chrono::milliseconds{10});
-
-        // Client 2 uploads two move instructions.
-        // The sync client uploads at most 128 KB of data so we make the first changeset large enough so two upload
-        // messages are sent to the server instead of one. Each change is transformed against the changes from
-        // Client 1.
-
-        // First change discards the first change (move(7, 2)) of Client 1.
-        move_element(db_2, 7, 0, 200 * 1024);
-        // Second change is tranformed against an empty reciprocal changeset as result of the change above.
-        move_element(db_2, 7, 5);
-        db_2_session = fixture.make_bound_session(2, db_2, 0, "/test");
-
-        db_1_session->wait_for_upload_complete_or_client_stopped();
-        db_2_session->wait_for_upload_complete_or_client_stopped();
-
-        db_1_session->wait_for_download_complete_or_client_stopped();
-        db_2_session->wait_for_download_complete_or_client_stopped();
-    }
-
-    ReadTransaction rt_1(db_1);
-    ReadTransaction rt_2(db_2);
-    const Group& group_1 = rt_1;
-    const Group& group_2 = rt_2;
-    group_1.verify();
-    group_2.verify();
-    CHECK(compare_groups(rt_1, rt_2));
+    VersionInfo version_info;
+    util::StderrLogger logger;
+    auto transact = db->start_read();
+    CHECK_THROW_EX(history.integrate_server_changesets({}, nullptr, util::Span(&server_changeset, 1), version_info,
+                                                       DownloadBatchState::SteadyState, logger, transact),
+                   sync::IntegrationException,
+                   StringData(e.what()).contains("Failed to parse received changeset: Invalid interned string"));
 }
 
 TEST(Sync_ServerVersionsSkippedFromDownloadCursor)
@@ -6969,5 +6773,205 @@ TEST(Sync_ServerVersionsSkippedFromDownloadCursor)
     CHECK_EQUAL(progress.upload.last_integrated_server_version,
                 expected_progress.upload.last_integrated_server_version);
 }
+
+TEST(Sync_NonIncreasingServerVersions)
+{
+    TEST_CLIENT_DB(db);
+
+    auto& history = get_history(db);
+    history.set_client_file_ident(SaltedFileIdent{2, 0x1234567812345678}, false);
+    timestamp_type timestamp{1};
+    history.set_local_origin_timestamp_source([&] {
+        return ++timestamp;
+    });
+
+    auto latest_local_version = [&] {
+        auto tr = db->start_write();
+        tr->add_table_with_primary_key("class_foo", type_String, "_id")->add_column(type_Int, "int_col");
+        return tr->commit();
+    }();
+
+    std::vector<Changeset> server_changesets;
+    auto prep_changeset = [&](auto pk_name, auto int_col_val) {
+        Changeset changeset;
+        changeset.version = 10;
+        changeset.last_integrated_remote_version = latest_local_version - 1;
+        changeset.origin_timestamp = ++timestamp;
+        changeset.origin_file_ident = 1;
+        instr::PrimaryKey pk{changeset.intern_string(pk_name)};
+        auto table_name = changeset.intern_string("foo");
+        auto col_name = changeset.intern_string("int_col");
+        instr::EraseObject erase_1;
+        erase_1.object = pk;
+        erase_1.table = table_name;
+        changeset.push_back(erase_1);
+        instr::CreateObject create_1;
+        create_1.object = pk;
+        create_1.table = table_name;
+        changeset.push_back(create_1);
+        instr::Update update_1;
+        update_1.table = table_name;
+        update_1.object = pk;
+        update_1.field = col_name;
+        update_1.value = instr::Payload{int64_t(int_col_val)};
+        changeset.push_back(update_1);
+        server_changesets.push_back(std::move(changeset));
+    };
+    prep_changeset("bizz", 1);
+    prep_changeset("buzz", 2);
+    prep_changeset("baz", 3);
+    prep_changeset("bar", 4);
+    ++server_changesets.back().version;
+
+    std::vector<ChangesetEncoder::Buffer> encoded;
+    std::vector<Transformer::RemoteChangeset> server_changesets_encoded;
+    for (const auto& changeset : server_changesets) {
+        encoded.emplace_back();
+        encode_changeset(changeset, encoded.back());
+        server_changesets_encoded.emplace_back(changeset.version, changeset.last_integrated_remote_version,
+                                               BinaryData(encoded.back().data(), encoded.back().size()),
+                                               changeset.origin_timestamp, changeset.origin_file_ident);
+    }
+
+    SyncProgress progress = {};
+    progress.download.server_version = server_changesets.back().version;
+    progress.download.last_integrated_client_version = latest_local_version - 1;
+    progress.latest_server_version.version = server_changesets.back().version;
+    progress.latest_server_version.salt = 0x7876543217654321;
+
+    uint_fast64_t downloadable_bytes = 0;
+    VersionInfo version_info;
+    util::StderrLogger logger;
+    auto transact = db->start_read();
+    history.integrate_server_changesets(progress, &downloadable_bytes, server_changesets_encoded, version_info,
+                                        DownloadBatchState::SteadyState, logger, transact);
+}
+
+
+TEST(Sync_DanglingLinksCountInPriorSize)
+{
+    SHARED_GROUP_TEST_PATH(path);
+    ClientReplication repl;
+    auto local_db = realm::DB::create(repl, path);
+    auto& history = repl.get_history();
+    history.set_client_file_ident(sync::SaltedFileIdent{1, 123456}, true);
+
+    version_type last_version, last_version_observed = 0;
+    auto dump_uploadable = [&] {
+        UploadCursor upload_cursor{last_version_observed, 0};
+        std::vector<sync::ClientHistory::UploadChangeset> changesets_to_upload;
+        version_type locked_server_version = 0;
+        history.find_uploadable_changesets(upload_cursor, last_version, changesets_to_upload, locked_server_version);
+        CHECK_EQUAL(changesets_to_upload.size(), static_cast<size_t>(1));
+        realm::sync::Changeset parsed_changeset;
+        auto unparsed_changeset = changesets_to_upload[0].changeset.get_first_chunk();
+        realm::util::SimpleInputStream changeset_stream(unparsed_changeset);
+        realm::sync::parse_changeset(changeset_stream, parsed_changeset);
+        test_context.logger->info("changeset at version %1: %2", last_version, parsed_changeset);
+        last_version_observed = last_version;
+        return parsed_changeset;
+    };
+
+    TableKey source_table_key, target_table_key;
+    {
+        auto wt = local_db->start_write();
+        auto source_table = wt->add_table_with_primary_key("class_source", type_String, "_id");
+        auto target_table = wt->add_table_with_primary_key("class_target", type_String, "_id");
+        source_table->add_column_list(*target_table, "links");
+
+        source_table_key = source_table->get_key();
+        target_table_key = target_table->get_key();
+
+        auto obj_to_keep = target_table->create_object_with_primary_key(std::string{"target1"});
+        auto obj_to_delete = target_table->create_object_with_primary_key(std::string{"target2"});
+        auto source_obj = source_table->create_object_with_primary_key(std::string{"source"});
+
+        auto links_list = source_obj.get_linklist("links");
+        links_list.add(obj_to_keep.get_key());
+        links_list.add(obj_to_delete.get_key());
+        last_version = wt->commit();
+    }
+
+    dump_uploadable();
+
+    {
+        // Simulate removing the object via the sync client so we get a dangling link
+        TempShortCircuitReplication disable_repl(repl);
+        auto wt = local_db->start_write();
+        auto target_table = wt->get_table(target_table_key);
+        auto obj = target_table->get_object_with_primary_key(std::string{"target2"});
+        obj.invalidate();
+        last_version = wt->commit();
+    }
+
+    {
+        auto wt = local_db->start_write();
+        auto source_table = wt->get_table(source_table_key);
+        auto target_table = wt->get_table(target_table_key);
+
+        auto obj_to_add = target_table->create_object_with_primary_key(std::string{"target3"});
+
+        auto source_obj = source_table->get_object_with_primary_key(std::string{"source"});
+        auto links_list = source_obj.get_linklist("links");
+        links_list.add(obj_to_add.get_key());
+        last_version = wt->commit();
+    }
+
+    auto changeset = dump_uploadable();
+    CHECK_EQUAL(changeset.size(), static_cast<size_t>(2));
+    auto changeset_it = changeset.end();
+    --changeset_it;
+    auto last_instr = *changeset_it;
+    CHECK_EQUAL(last_instr->type(), Instruction::Type::ArrayInsert);
+    auto arr_insert_instr = last_instr->get_as<Instruction::ArrayInsert>();
+    CHECK_EQUAL(changeset.get_string(arr_insert_instr.table), StringData("source"));
+    CHECK(arr_insert_instr.value.type == sync::instr::Payload::Type::Link);
+    CHECK_EQUAL(changeset.get_string(mpark::get<InternString>(arr_insert_instr.value.data.link.target)),
+                StringData("target3"));
+    CHECK_EQUAL(arr_insert_instr.prior_size, 2);
+}
+
+// This test calls row_for_object_id() for various object ids and tests that
+// the right value is returned including that no assertions are hit.
+TEST(Sync_RowForGlobalKey)
+{
+    TEST_CLIENT_DB(db);
+
+    {
+        WriteTransaction wt(db);
+        TableRef table = wt.add_table("class_foo");
+        table->add_column(type_Int, "i");
+        wt.commit();
+    }
+
+    // Check that various object_ids are not in the table.
+    {
+        ReadTransaction rt(db);
+        ConstTableRef table = rt.get_table("class_foo");
+        CHECK(table);
+
+        // Default constructed GlobalKey
+        {
+            GlobalKey object_id;
+            auto row_ndx = table->get_objkey(object_id);
+            CHECK_NOT(row_ndx);
+        }
+
+        // GlobalKey with small lo and hi values
+        {
+            GlobalKey object_id{12, 24};
+            auto row_ndx = table->get_objkey(object_id);
+            CHECK_NOT(row_ndx);
+        }
+
+        // GlobalKey with lo and hi values past the 32 bit limit.
+        {
+            GlobalKey object_id{uint_fast64_t(1) << 50, uint_fast64_t(1) << 52};
+            auto row_ndx = table->get_objkey(object_id);
+            CHECK_NOT(row_ndx);
+        }
+    }
+}
+
 
 } // unnamed namespace
